# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Omniport administration backend.",
        "title": "Omniport API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/master/menus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all menu records",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-menus"
                ],
                "summary": "Get all menus",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new menu record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-menus"
                ],
                "summary": "Create menu",
                "parameters": [
                    {
                        "description": "Menu payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menu.Menu"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/master/menus/search": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve menus with server-side pagination, filtering, sorting, and download range support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-menus"
                ],
                "summary": "Search menus",
                "parameters": [
                    {
                        "description": "Menu search payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menu.SearchMenusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.MetaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/master/menus/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing menu by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-menus"
                ],
                "summary": "Update menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Menu payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menu.Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete menu by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-menus"
                ],
                "summary": "Delete menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/master/references": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all reference headers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-references"
                ],
                "summary": "Get all references",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create or update reference header with details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-references"
                ],
                "summary": "Save reference",
                "parameters": [
                    {
                        "description": "Reference payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reference.PosmReference"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/master/references/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve reference header and details by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-references"
                ],
                "summary": "Get reference detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete reference by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master-references"
                ],
                "summary": "Delete reference",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "helper.DownloadQuery": {
            "type": "object",
            "properties": {
                "is_download": {
                    "type": "boolean"
                },
                "range_end": {
                    "type": "integer"
                },
                "range_start": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "helper.MetaResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "helper.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "helper.SortQuery": {
            "type": "object",
            "properties": {
                "by": {
                    "type": "string"
                },
                "order": {
                    "type": "string"
                }
            }
        },
        "menu.Menu": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "integer"
                },
                "creation_by": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_updated_by": {
                    "type": "string"
                },
                "last_updated_date": {
                    "type": "string"
                },
                "menu_code": {
                    "type": "string"
                },
                "menu_desc": {
                    "type": "string"
                },
                "menu_header_id": {
                    "type": "integer"
                },
                "menu_icon": {
                    "type": "string"
                },
                "menu_level": {
                    "type": "integer"
                },
                "menu_order": {
                    "type": "integer"
                },
                "menu_status": {
                    "type": "integer"
                },
                "menu_text": {
                    "type": "string"
                },
                "menu_url": {
                    "type": "string"
                },
                "parent_menu_id": {
                    "type": "integer"
                }
            }
        },
        "menu.SearchMenusRequest": {
            "type": "object",
            "properties": {
                "download": {
                    "$ref": "#/definitions/helper.DownloadQuery"
                },
                "filters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "sort": {
                    "$ref": "#/definitions/helper.SortQuery"
                }
            }
        },
        "reference.PosmReference": {
            "type": "object",
            "properties": {
                "branch_code": {
                    "type": "integer"
                },
                "creation_by": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reference.PosmReferenceD"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "id_ref_file": {
                    "type": "string"
                },
                "id_table": {
                    "type": "string"
                },
                "kd_aktif": {
                    "type": "string"
                },
                "keterangan": {
                    "type": "string"
                },
                "last_updated_by": {
                    "type": "string"
                },
                "last_updated_date": {
                    "type": "string"
                },
                "level_akses": {
                    "type": "string"
                },
                "program_name": {
                    "type": "string"
                },
                "ref1": {
                    "type": "string"
                },
                "terminal_code": {
                    "type": "integer"
                },
                "val1": {
                    "type": "integer"
                }
            }
        },
        "reference.PosmReferenceD": {
            "type": "object",
            "properties": {
                "branch_code": {
                    "type": "integer"
                },
                "creation_by": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_ref_file": {
                    "type": "string"
                },
                "id_ref_key": {
                    "type": "string"
                },
                "id_table": {
                    "type": "string"
                },
                "kd_aktif": {
                    "type": "string"
                },
                "ket_ref_data": {
                    "type": "string"
                },
                "last_updated_by": {
                    "type": "string"
                },
                "last_updated_date": {
                    "type": "string"
                },
                "level_akses": {
                    "type": "string"
                },
                "program_name": {
                    "type": "string"
                },
                "rc": {
                    "type": "integer"
                },
                "ref1": {
                    "type": "string"
                },
                "ref2": {
                    "type": "string"
                },
                "ref3": {
                    "type": "string"
                },
                "ref4": {
                    "type": "string"
                },
                "ref5": {
                    "type": "string"
                },
                "terminal_code": {
                    "type": "integer"
                },
                "val1": {
                    "type": "integer"
                },
                "val2": {
                    "type": "integer"
                },
                "val3": {
                    "type": "integer"
                },
                "val4": {
                    "type": "integer"
                },
                "val5": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}